(QENG-3276) Extract code into new Parser module
This commit moves the NODE_REGEX constant and handful of methods tightly
coupled to it into a new module called Parser. This module is
responsible for dealing with the raw user input string and turning it
into data structures that the Generator and hypervisors can operate on.
